Ported the code from the demo for npx (#3)
This reuses the code from the original demo, with no additional
improvements. The bin entry in `package.json` is used by NPX
when running the command, which points to a CLI script with a
shebang, which helps NPX run it.

The `bin` entry is a simple value as opposed to a key:value map,
so that If the package change names, this entry will be still valid, while
having a named entry requires us to update it as well in the event of a
renaming.