Draft of the token generation web app
This drafts the replacement of the generate-oidc-token CLI by a Web app.
The app is a NextJS app based on our template. The big difference is
that this app handles authentication on the server side, because the
browser side does not handle refresh tokens.

The interaction goes as follows:
- The user provides relevant information to the front-end. They submit.
- Submitting redirects the user to the backend app, which collects the
provided information, and uses it to redirect the user to the IdP.
- The user logs in to the IdP, and gets redirected back to the backend.
- The backend finishes authentication and gets a refresh token. It then
redirects the user to the frontend with all the information needed to
authenticate a script.
- The information is finally displayed to the user by the front end.

Note that until
https://github.com/inrupt/solid-client-authn-js/pull/1259 is merged and
published, this cannot work.